refactor: split ChatService into focused modules
Extract IntentClassifier, ResultCompactor, and ChatContextBuilder from
the 1700-line ChatService into separate files, reducing it to ~1060 lines.
No behavior changes — same public API (chat(), generatePlanForReview()).
