+++ mcp_backend/src/services/chat-service.ts
@@ -27,21 +27,20 @@ import {
 import { ModelSelector } from '@secondlayer/shared';
 import type { IEmbeddingPort, ILLMPort } from '../domain/ports/index.js';
 import {
-  CHAT_SYSTEM_PROMPT,
-  CHAT_INTENT_CLASSIFICATION_PROMPT,
-  DOMAIN_TOOL_MAP,
-  DEFAULT_TOOLS,
-  VALID_QUERY_TYPES,
   buildPlanGenerationMessages,
   ExecutionPlan,
   type QueryType,
   type ChatIntentClassification,
 } from '../prompts/chat-system-prompt.js';
-import { buildEnrichedSystemPrompt, SCENARIO_CATALOG } from '../prompts/tool-registry-catalog.js';
 import { QUERY_TYPE_CONFIG } from '../prompts/query-type-config.js';
 import { ChatSearchCacheService, isCourtSearchTool } from './chat-search-cache-service.js';
 import type { ShepardizationService, ShepardizationResult } from './shepardization-service.js';
 import { extractAllEvidence } from './evidence-extractor.js';
+
+import { BUDGET_LIMITS, CASE_NUMBER_REGEX, type BudgetKey } from './chat-constants.js';
+import { IntentClassifier } from './chat-intent-classifier.js';
+import { ResultCompactor } from './chat-result-compactor.js';
+import { ChatContextBuilder } from './chat-context-builder.js';
 
 // ============================
 // Types
@@ -76,19 +75,10 @@ interface PlanSession {
 }
 
 // ============================
-// Service
+// Constants
 // ============================
 
-// Budget-aware limits: deep analysis needs much more context
-const BUDGET_LIMITS = {
-  quick:    { maxResultChars: 6000,   maxContextChars: 48_000,  maxTokens: 4096,  maxToolCalls: 5,  resolutionSlice: 120 },
-  standard: { maxResultChars: 8000,   maxContextChars: 64_000,  maxTokens: 8192,  maxToolCalls: 7,  resolutionSlice: 300 },
-  deep:     { maxResultChars: 40_000, maxContextChars: 100_000, maxTokens: 16384, maxToolCalls: 20, resolutionSlice: 800 },
-} as const;
-type BudgetKey = keyof typeof BUDGET_LIMITS;
-
 const CITATION_CHECK_TIMEOUT_MS = 5_000;
-const CASE_NUMBER_REGEX = /\d+\/\d+\/\d{2,4}/g;
 
 // Tools where deduplication hashes only on the primary key (e.g. caseNumber),
 // ignoring secondary params like groupByInstance, includeFullText, maxDocs.
@@ -132,11 +122,17 @@ const DEPTH_OVERRIDES: Record<string, { standard: Record<string, any>; deep: Rec
 
 const PLAN_SESSION_TTL_MS = 5 * 60 * 1000; // 5 minutes
 
+// ============================
+// Service
+// ============================
+
 export class ChatService {
-  /** In-memory cache: conversationId → compressed summary of older history */
-  private historySummaryCache = new Map<string, { messageCount: number; summary: string }>();
   /** In-memory cache: planSessionId → classification + plan for two-phase flow */
   private planSessions = new Map<string, PlanSession>();
+
+  private intentClassifier: IntentClassifier;
+  private resultCompactor: ResultCompactor;
+  private contextBuilder: ChatContextBuilder;
 
   constructor(
     private toolRegistry: ToolRegistry,
@@ -148,6 +144,17 @@ export class ChatService {
     private shepardizationService?: ShepardizationService,
     private embeddingService?: IEmbeddingPort
   ) {
+    // Wire up cost recorder adapter for sub-modules
+    const costRecorder = {
+      recordStreamingCost: (requestId: string, provider: string, model: string, usage: any, task: string) => {
+        this.recordStreamingCost(requestId, provider as LLMProvider, model, usage, task);
+      },
+    };
+
+    this.intentClassifier = new IntentClassifier(toolRegistry, queryPlanner, llm, costRecorder);
+    this.resultCompactor = new ResultCompactor(embeddingService);
+    this.contextBuilder = new ChatContextBuilder(llm, costRecorder);
+
     // Periodic cleanup of expired plan sessions (every 2 minutes)
     setInterval(() => {
       const now = Date.now();
@@ -169,8 +176,8 @@ export class ChatService {
     userId?: string,
     requestId?: string
   ): Promise<{ plan: ExecutionPlan; planSessionId: string; queryType: QueryType } | null> {
-    const classification = await this.classifyChatIntent(query, requestId);
-    const toolDefs = await this.filterTools(classification.domains, classification.slots);
+    const classification = await this.intentClassifier.classify(query, requestId);
+    const toolDefs = await this.intentClassifier.filterTools(classification.domains, classification.slots);
     const plan = await this.generateExecutionPlan(query, classification, toolDefs, requestId);
 
     if (!plan) return null;
@@ -262,14 +269,14 @@ export class ChatService {
           logger.warn('[ChatService] Plan session expired or not found, regenerating', {
             planSessionId: request.planSessionId,
           });
-          classification = await this.classifyChatIntent(query, requestId);
-          toolDefs = await this.filterTools(classification.domains, classification.slots);
+          classification = await this.intentClassifier.classify(query, requestId);
+          toolDefs = await this.intentClassifier.filterTools(classification.domains, classification.slots);
           plan = await this.generateExecutionPlan(query, classification, toolDefs, requestId);
         }
       } else {
         // Standard flow: classify + generate plan
-        classification = await this.classifyChatIntent(query, requestId);
-        toolDefs = await this.filterTools(classification.domains, classification.slots);
+        classification = await this.intentClassifier.classify(query, requestId);
+        toolDefs = await this.intentClassifier.filterTools(classification.domains, classification.slots);
         plan = await this.generateExecutionPlan(query, classification, toolDefs, requestId);
       }
 
@@ -402,7 +409,7 @@ export class ChatService {
 
       // 5. Build messages with token-aware context window + injected plan
       const limits = BUDGET_LIMITS[effectiveBudget as BudgetKey] || BUDGET_LIMITS.standard;
-      const messages = await this.buildContextMessages(history, query, classification.domains, plan, limits.maxContextChars, request.conversationId, requestId, classification.queryType);
+      const messages = await this.contextBuilder.build(history, query, classification.domains, plan, limits.maxContextChars, request.conversationId, requestId, classification.queryType);
 
       // Log estimated prompt size for rate-limit debugging
       const totalChars = messages.reduce((sum, m) => sum + (m.content?.length || 0), 0);
@@ -616,7 +623,7 @@ export class ChatService {
 
           collectedThinkingSteps.push({ tool: call.name, params: call.arguments, result: toolResult });
 
-          const summarized = this.summarizeResult(toolResult, limits);
+          const summarized = this.resultCompactor.summarize(toolResult, limits);
 
           // Send the FULL result to the frontend for evidence extraction (decisions, citations).
           // The summarized version is only used for the LLM conversation to save tokens.
@@ -662,7 +669,7 @@ export class ChatService {
               tool: (m as any).tool_call_id || 'unknown',
               content: typeof m.content === 'string' ? m.content : JSON.stringify(m.content),
             }));
-            const compacted = await this.ragCompactToolResults(query, toolContents, limits.maxResultChars * 2);
+            const compacted = await this.resultCompactor.ragCompact(query, toolContents, limits.maxResultChars * 2);
             // Replace tool message contents with compacted versions
             let compIdx = 0;
             for (const msg of messages) {
@@ -812,365 +819,6 @@ export class ChatService {
   }
 
   /**
-   * Token-aware sliding window with history compression.
-   * Recent messages (last 4) are kept verbatim; older messages are LLM-summarized.
-   */
-  private async buildContextMessages(
-    history: Array<{ role: 'user' | 'assistant'; content: string }>,
-    query: string,
-    classifiedDomains?: string[],
-    plan?: ExecutionPlan,
-    maxContextChars?: number,
-    conversationId?: string,
-    requestId?: string,
-    queryType?: QueryType
-  ): Promise<UnifiedMessage[]> {
-    const contextBudget = maxContextChars || BUDGET_LIMITS.standard.maxContextChars;
-
-    // Get preferred scenarios from queryType config
-    const qtConfig = queryType ? QUERY_TYPE_CONFIG[queryType] : undefined;
-    const preferredScenarioIds = qtConfig?.preferredScenarios;
-
-    let enrichedPrompt = buildEnrichedSystemPrompt(
-      CHAT_SYSTEM_PROMPT,
-      SCENARIO_CATALOG,
-      classifiedDomains,
-      preferredScenarioIds
-    );
-
-    // Inject grounding note from queryType config
-    if (qtConfig?.groundingNote) {
-      enrichedPrompt += `\n\n## Обмеження відповіді (queryType: ${queryType})\n${qtConfig.groundingNote}`;
-    }
-
-    // Inject execution plan into system prompt
-    if (plan) {
-      const stepsText = plan.steps
-        .map((s) => {
-          const paramsStr = JSON.stringify(s.params);
-          const depsStr = s.depends_on?.length ? ` (після кроків ${s.depends_on.join(', ')})` : '';
-          return `Крок ${s.id}: ${s.tool}(${paramsStr})${depsStr}\n  → Мета: ${s.purpose}`;
-        })
-        .join('\n\n');
-
-      enrichedPrompt += `\n\n## План виконання
-
-Ціль: ${plan.goal}
-
-${stepsText}
-
-ВАЖЛИВО: Виконай кроки в зазначеному порядку. Можеш адаптувати параметри на основі результатів попередніх кроків.
-Використовуй РЕАЛЬНІ дані з результатів інструментів. НІКОЛИ не залишай плейсхолдери — тільки конкретні факти.
-Якщо інструмент не повернув певну інформацію — напиши "інформація не знайдена".`;
-    }
-
-    const messages: UnifiedMessage[] = [
-      { role: 'system', content: enrichedPrompt },
-    ];
-
-    const systemChars = enrichedPrompt.length;
-    const queryChars = query.length;
-    let availableChars = contextBudget - systemChars - queryChars;
-
-    // Split history: last 4 messages are "recent", the rest are "older"
-    const RECENT_COUNT = 4;
-    const recentHistory = history.slice(-RECENT_COUNT);
-    const olderHistory = history.slice(0, Math.max(0, history.length - RECENT_COUNT));
-
-    // Compress older history into a summary if it exists
-    if (olderHistory.length > 0) {
-      const summary = await this.compressOlderHistory(olderHistory, conversationId, requestId);
-      if (summary) {
-        const summaryChars = summary.length + 20;
-        if (availableChars - summaryChars > 0) {
-          messages.push({ role: 'user', content: summary });
-          availableChars -= summaryChars;
-        }
-      }
-    }
-
-    // Add recent messages verbatim (truncated to 2000 chars each), from most recent backwards
-    const recentMessages: UnifiedMessage[] = [];
-    for (let i = recentHistory.length - 1; i >= 0; i--) {
-      const content = recentHistory[i].content.slice(0, 2000);
-      const msgChars = content.length + 20;
-      if (availableChars - msgChars < 0) break;
-      availableChars -= msgChars;
-      recentMessages.unshift({ role: recentHistory[i].role, content });
-    }
-
-    messages.push(...recentMessages);
-    messages.push({ role: 'user', content: query });
-    return messages;
-  }
-
-  /**
-   * Compress older chat history into a concise summary using a quick LLM call.
-   * Caches summaries per conversation to avoid re-summarizing on each iteration.
-   */
-  private async compressOlderHistory(
-    olderMessages: Array<{ role: string; content: string }>,
-    conversationId?: string,
-    requestId?: string
-  ): Promise<string | null> {
-    // If total content is small, just concatenate
-    const totalChars = olderMessages.reduce((sum, m) => sum + m.content.length, 0);
-    if (totalChars < 2000) {
-      const concat = olderMessages.map(m => `${m.role}: ${m.content}`).join('\n');
-      return `[Контекст попередніх повідомлень]: ${concat}`;
-    }
-
-    // Check cache
-    if (conversationId) {
-      const cached = this.historySummaryCache.get(conversationId);
-      if (cached && cached.messageCount === olderMessages.length) {
-        return cached.summary;
-      }
-    }
-
-    try {
-      const llm = this.llm;
-
-      const historyText = olderMessages
-        .map(m => `${m.role}: ${m.content.slice(0, 1000)}`)
-        .join('\n---\n');
-
-      const response = await llm.chatCompletion(
-        {
-          messages: [
-            {
-              role: 'system',
-              content: 'Стисло підсумуй контекст розмови (до 200 слів). Збережи ключові юридичні факти, номери справ, посилання на закони та прийняті рішення. Відповідай українською.',
-            },
-            { role: 'user', content: historyText },
-          ],
-          max_tokens: 400,
-          temperature: 0.1,
-        },
-        'quick'
-      );
-
-      if (requestId && response.usage) {
-        this.recordStreamingCost(requestId, response.provider, response.model, response.usage, 'history_compression');
-      }
-
-      const summary = `[Контекст попередніх повідомлень]: ${response.content || ''}`;
-
-      // Cache the summary
-      if (conversationId) {
-        this.historySummaryCache.set(conversationId, {
-          messageCount: olderMessages.length,
-          summary,
-        });
-      }
-
-      logger.info('[ChatService] Compressed older history', {
-        originalMessages: olderMessages.length,
-        originalChars: totalChars,
-        summaryChars: summary.length,
-      });
-
-      return summary;
-    } catch (err: any) {
-      logger.warn('[ChatService] History compression failed, using truncated concat', { error: err.message });
-      // Fallback: truncated concatenation
-      const concat = olderMessages
-        .map(m => `${m.role}: ${m.content.slice(0, 300)}`)
-        .join('\n');
-      return `[Контекст попередніх повідомлень]: ${concat}`.slice(0, 3000);
-    }
-  }
-
-  /**
-   * Classify chat intent using a fast LLM call (gpt-4o-mini ~200ms).
-   * Falls back to keyword-based QueryPlanner on error.
-   */
-  private async classifyChatIntent(query: string, requestId?: string): Promise<ChatIntentClassification> {
-    try {
-      const llm = this.llm;
-
-      const classifyChars = CHAT_INTENT_CLASSIFICATION_PROMPT.length + query.length;
-      logger.debug('[ChatService] Intent classification prompt size', {
-        chars: classifyChars,
-        estimatedTokens: Math.ceil(classifyChars / 3.5),
-      });
-
-      const response = await llm.chatCompletion(
-        {
-          messages: [
-            { role: 'system', content: CHAT_INTENT_CLASSIFICATION_PROMPT },
-            { role: 'user', content: query },
-          ],
-          max_tokens: 300,
-          temperature: 0.1,
-          response_format: { type: 'json_object' },
-        },
-        'quick'
-      );
-
-      // Record classification LLM cost
-      if (requestId && response.usage) {
-        this.recordStreamingCost(requestId, response.provider, response.model, response.usage, 'intent_classification');
-      }
-
-      const content = response.content || '{}';
-
-      // Extract JSON from response (may be wrapped in markdown code blocks)
-      const jsonMatch = content.match(/\{[\s\S]*\}/);
-      if (!jsonMatch) {
-        throw new Error('No JSON found in classification response');
-      }
-
-      const parsed = JSON.parse(jsonMatch[0]);
-
-      const domains = Array.isArray(parsed.domains) && parsed.domains.length > 0
-        ? parsed.domains
-        : ['court'];
-      const keywords = typeof parsed.keywords === 'string' ? parsed.keywords : query;
-      let slots: Record<string, any> | undefined = parsed.slots && typeof parsed.slots === 'object' && Object.keys(parsed.slots).length > 0
-        ? parsed.slots
-        : undefined;
-
-      // Force-include domains based on extracted slots
-      if (slots?.edrpou && !domains.includes('registry')) {
-        domains.push('registry');
-      }
-      if (slots?.case_number && !domains.includes('court')) {
-        domains.push('court');
-      }
-      if (slots?.law_reference && !domains.includes('legislation')) {
-        domains.push('legislation');
-      }
-
-      // Safety net: extract case_number from query if LLM missed it
-      if (!slots?.case_number) {
-        const caseMatch = query.match(/\d+\/\d+\/\d{2,4}/);
-        if (caseMatch) {
-          if (!slots) slots = {};
-          slots.case_number = caseMatch[0];
-        }
-      }
-
-      // Keyword-based safety net for registry queries
-      const lowerQuery = query.toLowerCase();
-      const registryKeywords = ['тов ', 'тов "', 'тов «', 'фоп ', 'пп ', 'ат ', 'єдрпоу', 'edrpou', 'підприємство', 'компанія', 'юридична особа'];
-      const hasRegistryKeyword = registryKeywords.some(kw => lowerQuery.includes(kw));
-      const hasEdrpouPattern = /\b\d{8}\b/.test(query);
-      if ((hasRegistryKeyword || hasEdrpouPattern) && !domains.includes('registry')) {
-        domains.push('registry');
-      }
-
-      // Safety net for documents/vault queries
-      const documentsKeywords = ['vault', 'сховищ', 'завантажив', 'завантажені', 'мої документи', 'мої файли', 'загрузил', 'зарузил', 'загруженн', 'зберіг', 'збережені', 'uploaded', 'my documents', 'my files'];
-      const hasDocumentsKeyword = documentsKeywords.some(kw => lowerQuery.includes(kw));
-      if (hasDocumentsKeyword && !domains.includes('documents')) {
-        domains.push('documents');
-      }
-
-      // Parse queryType with validation and safety coercions
-      let queryType: QueryType = 'legal_consultation';
-      if (parsed.queryType && VALID_QUERY_TYPES.has(parsed.queryType)) {
-        queryType = parsed.queryType as QueryType;
-      }
-
-      // Safety coercions based on slots
-      if (slots?.case_number && queryType === 'legal_consultation') {
-        queryType = 'case_lookup';
-      }
-      if (slots?.edrpou && queryType === 'legal_consultation') {
-        queryType = 'registry_lookup';
-      }
-      if (slots?.law_reference && queryType === 'legal_consultation') {
-        queryType = 'legislation_lookup';
-      }
-
-      // Regex-based queryType coercions when LLM defaults to legal_consultation
-      if (queryType === 'legal_consultation') {
-        // Legislation: "ст. 16 ЦК", "стаття 382 КК", "ч. 3 ст. 16"
-        if (/ст(?:атт[яію])?\s*\.?\s*\d+/i.test(query) || /(?:^|\s)(?:ЦК|КК|ГПК|ЦПК|КАС|ГК|ЗК|СК|КЗпП|цк|кк|гпк|цпк|кас|гк|зк|ск|кзпп)(?:\s|$|,|\.)/.test(query)) {
-          queryType = 'legislation_lookup';
-        }
-        // Registry: ЄДРПОУ, 8-digit code, ТОВ/ПАТ/ФОП lookup
-        else if (hasEdrpouPattern || lowerQuery.includes('єдрпоу') || lowerQuery.includes('edrpou')) {
-          queryType = 'registry_lookup';
-        }
-        // Practice analysis: "аналіз практики", "судова практика", "як суди"
-        else if (/проаналізу|аналіз практик|судова практика|знайти справи|знайти практику|огляд практики|яка практика|як суди|позиція судів/i.test(query)) {
-          queryType = 'practice_analysis';
-        }
-        // Document drafting: "напиши позовну", "зразок скарги", "склади заяву"
-        else if (/\b(?:напиш[иі]|склад[иі]|підготуй|зразок|шаблон)\b.*\b(?:позов|скарг|заяв|клопотан|претенз)/i.test(query)) {
-          queryType = 'document_drafting';
-        }
-        // Due diligence: "перевір контрагента", "due diligence"
-        else if (/due.?diligence|перевір.*контрагент|комплексна перевірка/i.test(query)) {
-          queryType = 'due_diligence';
-        }
-        // Parliament: "депутат", "законопроєкт", "голосування"
-        else if (domains.includes('parliament')) {
-          queryType = 'parliament_query';
-        }
-        // Documents/vault
-        else if (hasDocumentsKeyword || domains.includes('documents')) {
-          queryType = 'document_query';
-        }
-        // Comparative analysis: "негаторний чи віндикаційний", "який спосіб захисту"
-        else if (/\bчи\b.*\bпозов|який спосіб захисту|порівн.*підход|яка стаття підходить/i.test(query)) {
-          queryType = 'comparative_analysis';
-        }
-        // Unsupported: non-legal queries
-        else if (/(?:погод[аиу]|рецепт|футбол|спорт|кіно|фільм|музик|пісн)/i.test(lowerQuery) && !domains.some((d: string) => d !== 'court')) {
-          queryType = 'unsupported';
-        }
-        // Unsupported: aggregated judge statistics
-        else if (/рейтинг.*(судд|суд)|статистик.*(судд|суд)|відсот.*(задовол|відмов).*судд|тенденці.*(закрит|судд)/i.test(lowerQuery)) {
-          queryType = 'unsupported';
-        }
-      }
-
-      let unsupportedReason = queryType === 'unsupported' && typeof parsed.unsupportedReason === 'string'
-        ? parsed.unsupportedReason
-        : undefined;
-
-      // Generate unsupportedReason if not provided by LLM
-      if (queryType === 'unsupported' && !unsupportedReason) {
-        if (/рейтинг.*(судд|суд)|статистик.*(судд|суд)|відсот.*(задовол|відмов)|тенденці.*(закрит|судд)/i.test(lowerQuery)) {
-          unsupportedReason = 'Система не може агрегувати статистику по суддях (відсоток задоволених позовів, тенденції). Можу знайти конкретні справи за іменем судді або за темою.';
-        } else {
-          unsupportedReason = 'Цей запит виходить за межі можливостей юридичної системи SecondLayer. Я спеціалізуюся на українському праві: судова практика, законодавство, реєстри, парламентські дані.';
-        }
-      }
-
-      logger.info('[ChatService] LLM intent classification', { domains, keywords, slots, queryType, unsupportedReason });
-
-      return { domains, keywords, slots, queryType, unsupportedReason };
-    } catch (err: any) {
-      logger.warn('[ChatService] LLM classification failed, falling back to keyword matching', {
-        error: err.message,
-      });
-
-      // Fallback to keyword-based classification
-      const intent = await this.queryPlanner.classifyIntent(query, 'quick');
-      const fallbackSlots = (intent.slots as Record<string, any>) || {};
-
-      // Extract case_number from query if QueryPlanner didn't
-      if (!fallbackSlots.case_number) {
-        const caseMatch = query.match(/\d+\/\d+\/\d{2,4}/);
-        if (caseMatch) {
-          fallbackSlots.case_number = caseMatch[0];
-        }
-      }
-
-      return {
-        domains: intent.domains,
-        keywords: query,
-        slots: Object.keys(fallbackSlots).length > 0 ? fallbackSlots : undefined,
-        queryType: 'legal_consultation' as QueryType,
-      };
-    }
-  }
-
-  /**
    * Generate a structured execution plan: which tools to call, in what order,
    * with what parameters. Uses a fast LLM call (quick budget, ~200-400ms).
    * Falls back to undefined on error → agentic loop runs without a plan.
@@ -1287,47 +935,6 @@ ${stepsText}
   }
 
   /**
-   * Filter 45+ tools to a relevant subset based on intent domains.
-   */
-  private async filterTools(domains: string[], slots?: Record<string, any>): Promise<ToolDefinition[]> {
-    const allDefs = await this.toolRegistry.getAllToolDefinitions();
-
-    // Collect tool names from matching domains
-    const relevantNames = new Set<string>(DEFAULT_TOOLS);
-    for (const domain of domains) {
-      const mapped = DOMAIN_TOOL_MAP[domain];
-      if (mapped) {
-        for (const name of mapped) {
-          relevantNames.add(name);
-        }
-      }
-    }
-
-    // If EDRPOU is in slots, ensure all registry tools are included
-    if (slots?.edrpou) {
-      const registryTools = DOMAIN_TOOL_MAP.registry || [];
-      for (const name of registryTools) {
-        relevantNames.add(name);
-      }
-    }
-
-    // If no domains matched, use a broader set
-    if (relevantNames.size <= DEFAULT_TOOLS.length && domains.length > 0) {
-      // Add all tools from the 'legal_advice' domain as fallback
-      const fallback = DOMAIN_TOOL_MAP.legal_advice || [];
-      for (const name of fallback) {
-        relevantNames.add(name);
-      }
-    }
-
-    // Filter to only tools that actually exist in the registry
-    const filtered = allDefs.filter((d) => relevantNames.has(d.name));
-
-    // Cap at 15 tools — modern LLMs handle 15-20 tools fine
-    return filtered.slice(0, 15);
-  }
-
-  /**
    * Convert ToolRegistry definitions to LLM function calling format.
    */
   private convertToolDefs(defs: ToolDefinition[]): ToolDefinitionParam[] {
@@ -1336,254 +943,6 @@ ${stepsText}
       description: d.description,
       parameters: d.inputSchema || { type: 'object', properties: {} },
     }));
-  }
-
-  /**
-   * Summarize large tool results to prevent context window overflow.
-   * Budget-aware: deep analysis preserves much more content.
-   */
-  private summarizeResult(result: any, limits: typeof BUDGET_LIMITS[BudgetKey]): any {
-    if (!result) return { empty: true };
-
-    const maxLen = limits.maxResultChars;
-
-    // For MCP tool results with content array — try compact extraction first
-    if (result.content && Array.isArray(result.content)) {
-      const compacted = this.compactCourtResult(result, limits);
-      if (compacted) return compacted;
-
-      // Fallback: truncate text blocks
-      const cloned = { ...result, content: result.content.map((b: any) => ({ ...b })) };
-      for (const block of cloned.content) {
-        if (block.type === 'text' && typeof block.text === 'string') {
-          if (block.text.length > maxLen) {
-            block.text = block.text.slice(0, maxLen) + '\n\n[... результат скорочено]';
-          }
-        }
-      }
-      return cloned;
-    }
-
-    const text = typeof result === 'string' ? result : JSON.stringify(result);
-
-    if (text.length <= maxLen) {
-      return result;
-    }
-
-    // Generic truncation
-    return {
-      summary: text.slice(0, maxLen),
-      truncated: true,
-      original_length: text.length,
-    };
-  }
-
-  /**
-   * Compact court document chain/search results for LLM context.
-   * Never sends full_text to LLM — extracts key sections (FACTS, REASONING, DECISION) instead.
-   * Budget-aware section limits: quick=500, standard=1500, deep=3000 chars per section.
-   */
-  private compactCourtResult(result: any, limits: typeof BUDGET_LIMITS[BudgetKey]): any | null {
-    if (!result.content?.[0]?.text) return null;
-
-    let parsed: any;
-    try {
-      parsed = JSON.parse(result.content[0].text);
-    } catch {
-      return null;
-    }
-
-    const resSlice = limits.resolutionSlice;
-    const sectionLimit = limits.maxTokens > 4096 ? 3000 : (limits.maxResultChars <= 6000 ? 500 : 1500);
-
-    // Case documents chain: { case_number, total_documents, grouped_documents }
-    // For chain results we send ONLY metadata (no key_sections) so that ALL documents
-    // fit within the budget.  The LLM can call load_full_texts for deeper analysis.
-    if (parsed.grouped_documents && parsed.total_documents) {
-      const compact: any = {
-        case_number: parsed.case_number,
-        total_documents: parsed.total_documents,
-        grouped_documents: {},
-      };
-
-      for (const [instance, docs] of Object.entries(parsed.grouped_documents)) {
-        compact.grouped_documents[instance] = (docs as any[]).map((d: any) => {
-          const entry: any = {
-            doc_id: d.doc_id,
-            url: d.doc_id ? `https://reyestr.court.gov.ua/Review/${d.doc_id}` : undefined,
-            case_number: d.cause_num || d.case_number,
-            document_type: d.document_type,
-            instance: d.instance,
-            court: d.court,
-            judge: d.judge,
-            date: d.date,
-            resolution: d.resolution ? d.resolution.slice(0, resSlice) : undefined,
-          };
-          // Only include snippets for chain results — key_sections are too large
-          // and cause truncation when there are many documents (e.g. 29 docs × 3 sections × 3000 chars)
-          if (d.snippets) entry.snippets = d.snippets;
-          return entry;
-        });
-      }
-
-      const compactText = JSON.stringify(compact);
-      if (compactText.length > limits.maxResultChars) {
-        return { content: [{ type: 'text', text: compactText.slice(0, limits.maxResultChars) + '\n[... скорочено]' }] };
-      }
-      return { content: [{ type: 'text', text: compactText }] };
-    }
-
-    // Search results: { results: [...], total_count }
-    if (Array.isArray(parsed.results)) {
-      const compact = {
-        total_count: parsed.total_count || parsed.results.length,
-        results: parsed.results.map((r: any) => {
-          const entry: any = {
-            doc_id: r.doc_id,
-            url: r.doc_id ? `https://reyestr.court.gov.ua/Review/${r.doc_id}` : undefined,
-            case_number: r.cause_num || r.case_number,
-            document_type: r.document_type,
-            court: r.court_code || r.court,
-            judge: r.judge,
-            date: r.date || r.adjudication_date,
-            instance: r.instance,
-            resolution: r.resolution ? r.resolution.slice(0, resSlice) : undefined,
-          };
-          entry.key_sections = this.extractKeySections(r.sections, r.full_text, sectionLimit);
-          if (r.snippets) entry.snippets = r.snippets;
-          return entry;
-        }),
-      };
-
-      const compactText = JSON.stringify(compact);
-      if (compactText.length > limits.maxResultChars) {
-        return { content: [{ type: 'text', text: compactText.slice(0, limits.maxResultChars) + '\n[... скорочено]' }] };
-      }
-      return { content: [{ type: 'text', text: compactText }] };
-    }
-
-    return null;
-  }
-
-  /**
-   * Extract key sections (FACTS, COURT_REASONING, DECISION) from court document.
-   * Uses structured sections if available, otherwise falls back to regex extraction from full_text.
-   */
-  private extractKeySections(
-    sections: any,
-    fullText: string | undefined,
-    charLimit: number
-  ): { facts?: string; reasoning?: string; decision?: string } | undefined {
-    const KEY_SECTION_TYPES = ['FACTS', 'COURT_REASONING', 'DECISION', 'ВСТАНОВИВ', 'МОТИВУВАЛЬНА', 'РЕЗОЛЮТИВНА'];
-
-    const result: { facts?: string; reasoning?: string; decision?: string } = {};
-
-    // Try structured sections first
-    if (sections && Array.isArray(sections)) {
-      for (const s of sections) {
-        const type = (s.type || s.section_type || '').toUpperCase();
-        const text = s.text || s.content || '';
-        if (!text) continue;
-
-        if (type.includes('FACT') || type.includes('ВСТАНОВИВ')) {
-          result.facts = text.slice(0, charLimit);
-        } else if (type.includes('REASON') || type.includes('МОТИВ')) {
-          result.reasoning = text.slice(0, charLimit);
-        } else if (type.includes('DECISION') || type.includes('РЕЗОЛЮТ')) {
-          result.decision = text.slice(0, charLimit);
-        }
-      }
-    }
-
-    // If we didn't find sections from structured data, try regex from full_text
-    if (!result.facts && !result.reasoning && !result.decision && fullText) {
-      // Ukrainian court decisions commonly have: ВСТАНОВИВ, МОТИВУВАЛЬНА ЧАСТИНА, ВИРІШИВ/УХВАЛИВ/ПОСТАНОВИВ
-      const factsMatch = fullText.match(/ВСТАНОВИВ[:\s]*([\s\S]{10,}?)(?=(?:МОТИВУВАЛЬНА|ВИРІШИВ|УХВАЛИВ|ПОСТАНОВИВ)|$)/i);
-      const reasoningMatch = fullText.match(/МОТИВУВАЛЬНА[^:]*:[:\s]*([\s\S]{10,}?)(?=(?:ВИРІШИВ|УХВАЛИВ|ПОСТАНОВИВ|РЕЗОЛЮТИВНА)|$)/i);
-      const decisionMatch = fullText.match(/(?:ВИРІШИВ|УХВАЛИВ|ПОСТАНОВИВ)[:\s]*([\s\S]{10,}?)$/i);
-
-      if (factsMatch) result.facts = factsMatch[1].trim().slice(0, charLimit);
-      if (reasoningMatch) result.reasoning = reasoningMatch[1].trim().slice(0, charLimit);
-      if (decisionMatch) result.decision = decisionMatch[1].trim().slice(0, charLimit);
-    }
-
-    // Return undefined if no sections extracted to avoid empty object in JSON
-    return (result.facts || result.reasoning || result.decision) ? result : undefined;
-  }
-
-  /**
-   * RAG-based compaction of tool results when context gets too large.
-   * Embeds the user query, computes cosine similarity with each tool result,
-   * and keeps the top-K most relevant results fully while summarizing the rest.
-   * Only triggers when there are 3+ results and total chars exceed threshold.
-   */
-  private async ragCompactToolResults(
-    query: string,
-    toolResults: Array<{ tool: string; content: string }>,
-    maxChars: number
-  ): Promise<Array<{ tool: string; content: string }>> {
-    if (!this.embeddingService || toolResults.length < 3) return toolResults;
-
-    const totalChars = toolResults.reduce((sum, r) => sum + r.content.length, 0);
-    if (totalChars <= maxChars * 1.5) return toolResults;
-
-    try {
-      // Embed the query
-      const queryEmbedding = await this.embeddingService.generateEmbedding(query);
-
-      // Embed each tool result (use first 500 chars as representative)
-      const scored = await Promise.all(
-        toolResults.map(async (r) => {
-          const snippet = r.content.slice(0, 500);
-          const embedding = await this.embeddingService!.generateEmbedding(snippet);
-          const similarity = this.cosineSimilarity(queryEmbedding, embedding);
-          return { ...r, similarity };
-        })
-      );
-
-      // Sort by relevance (highest first)
-      scored.sort((a, b) => b.similarity - a.similarity);
-
-      // Keep top results fully until we approach the budget; summarize the rest
-      const compacted: Array<{ tool: string; content: string }> = [];
-      let usedChars = 0;
-
-      for (const item of scored) {
-        if (usedChars + item.content.length <= maxChars) {
-          compacted.push({ tool: item.tool, content: item.content });
-          usedChars += item.content.length;
-        } else {
-          // Summarize to just metadata line
-          const summary = `[${item.tool}]: результат скорочено (релевантність: ${item.similarity.toFixed(2)}, ${item.content.length} символів)`;
-          compacted.push({ tool: item.tool, content: summary });
-          usedChars += summary.length;
-        }
-      }
-
-      logger.info('[ChatService] RAG compacted tool results', {
-        originalResults: toolResults.length,
-        originalChars: totalChars,
-        compactedChars: usedChars,
-        fullResults: compacted.filter(r => !r.content.startsWith('[')).length,
-      });
-
-      return compacted;
-    } catch (err: any) {
-      logger.warn('[ChatService] RAG compaction failed, returning original results', { error: err.message });
-      return toolResults;
-    }
-  }
-
-  private cosineSimilarity(a: number[], b: number[]): number {
-    if (a.length !== b.length) return 0;
-    let dot = 0, normA = 0, normB = 0;
-    for (let i = 0; i < a.length; i++) {
-      dot += a[i] * b[i];
-      normA += a[i] * a[i];
-      normB += b[i] * b[i];
-    }
-    const denom = Math.sqrt(normA) * Math.sqrt(normB);
-    return denom === 0 ? 0 : dot / denom;
   }
 
   /**
